Refactoring code base to utilize better encapsulation
